+++ src/modules/assets/copy/modulesCopy.ts
@@ -39,12 +39,14 @@ export const modulesCopy = {
     proposalActionsFooter: {
         collapse: 'Collapse all',
         expand: 'Expand all',
+        more: 'More',
     },
     proposalActionsItem: {
         dropdownLabel: 'More',
         nativeSendAlert: 'Proceed with caution',
         nativeSendDescription: (amount: string, symbol: string) =>
             `This action attempts to send ${amount} ${symbol}. This could cause the action to fail or result in a loss of funds.`,
+        of: 'of',
         menu: {
             BASIC: 'Basic',
             dropdownLabel: 'View action as',